Feedback Analyzer/requirements.txt
Flask
beautifulsoup4
requests
python-dotenv
google-api-python-client